GH-4209: From now on, Theia is not shipped with Git.
Closes: #4209.
